+++ src/util/graph-templates/graph-helpers/resizeSvg.js
@@ -1,8 +1,6 @@
-export function resize(selection) {
+export default function resizeSvg(selection) {
   selection
     .transition()
     .attr('height', this.padding.top + this.height + this.padding.bottom + this.sourceHeight)
     .attr('width', this.padding.left + this.width + this.padding.right);
 }
-
-export function foo() {}