+++ src/util/graph-templates/graph-helpers/rowHelpers/createBars.js
@@ -1,0 +1,7 @@
+export default function(selection) {
+  return selection
+    .selectAll('rect.bar')
+    .data(d => d.values)
+    .join('rect')
+    .attr('class', 'bar');
+}